Move local-cluster executable code to the executable stanza
Remove executable code indirection in local-cluster exe code

And move CLI options back in the lib
